Added FX to firing, hitting enemies, and jumping
The FX on firing is a placeholder, and shares the same sprite as the one used on the enemies "got hit" FX. The FX is pretty simple and fade after about half a second, as I'm not super experienced with making visuals. Also, I tried and failed to fix the fact that enemies don't actually look towards the player, but rather at a corner of the screen.